LED/requirements.txt
opencv_python
glog
numpy
scikit_image
easydict
matplotlib
skimage
tensorflow_gpu==1.15.0
